Fix clang-tidy warnings

+++ src/sensesp/net/http_authenticator.cpp
@@ -6,7 +6,7 @@ String get_random_hex_string() {
   char buffer[33];  // buffer to hold 32 Hex Digit + /0
   int i;
   for (i = 0; i < 4; i++) {
-    sprintf(buffer + (i * 8), "%08x", esp_random());
+    sprintf(buffer + (i * 8), "%08x", static_cast<unsigned int>(esp_random()));
   }
   return String(buffer);
 }
